app/processing/image_processor.py
"""Extract text and generate descriptions from images"""
import logging
from io import BytesIO
from typing import Optional
from PIL import Image

logger = logging.getLogger(__name__)


class ImageProcessor:
    """Extract text and generate descriptions from images"""
    
    def __init__(self):
        pass
    
    async def process_image(self, image_source: str) -> dict:
        """Process image from URL or file path"""
        # Load image
        if image_source.startswith('http'):
            import requests
            response = requests.get(image_source)
            image = Image.open(BytesIO(response.content))
        else:
            image = Image.open(image_source)
        
        # Extract text (OCR)
        text = await self.ocr(image)
        
        # Generate description (if using vision model)
        description = await self.describe_image(image)
        
        return {
            "text": text,
            "description": description,
            "size": image.size,
            "format": image.format
        }
    
    async def ocr(self, image: Image.Image) -> str:
        """Extract text from image using OCR"""
        try:
            import pytesseract
            text = pytesseract.image_to_string(image)
            return text.strip()
        except ImportError:
            logger.warning("pytesseract not available for OCR")
            return ""
        except Exception as e:
            logger.error(f"OCR failed: {e}")
            return ""
    
    async def ocr_from_bytes(self, image_bytes: bytes) -> str:
        """OCR from image bytes"""
        image = Image.open(BytesIO(image_bytes))
        return await self.ocr(image)
    
    async def describe_image(self, image: Image.Image) -> str:
        """Generate image description using vision model"""
        # If you have access to GPT-4V or similar
        # Convert image to base64 and send to vision API
        # For now, return placeholder
        return "Image description feature requires vision model"

