requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Embedding
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector store
faiss-cpu>=1.7.4
# or faiss-gpu for GPU support

# Document processing
pypdf>=3.17.0
pymupdf>=1.23.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Audio processing
openai-whisper>=20231117

# Image processing
pytesseract>=0.3.10
pillow>=10.1.0

# Web scraping
playwright>=1.40.0
requests>=2.31.0

# Retrieval
rank-bm25>=0.2.2

# LLM clients
openai>=1.3.0
anthropic>=0.7.0

# Machine Learning
scikit-learn>=1.3.0

# Security
pyjwt>=2.8.0

# Monitoring
prometheus-client>=0.19.0
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
numpy>=1.24.0

